requirements.txt

faiss-cpu
numpy
